+++ src/hooks/useFormContext.tsx
@@ -1,8 +1,7 @@
-import React, {useContext} from 'react';
-import {FormInstance} from './useForm';
+import React, {Context, useContext} from 'react';
+import {FormInstance, Store} from './formInstance';
 
-export const FormContext = React.createContext({});
-
-const useFormContext = () => useContext(FormContext);
-
-export default useFormContext;
+export const FormContext = React.createContext({} as FormInstance);
+export function useFormContext<T = Store>() {
+  return useContext<FormInstance<T>>(FormContext as Context<FormInstance<T>>);
+}
